報酬、希望年収をIntegerからDecimalに変更、匿名解除後アイコンを表示していた箇所を非表示に変更

+++ lib/bright/recruitment_stock_users/recruitment_stock_user.ex
@@ -11,7 +11,7 @@ defmodule Bright.RecruitmentStockUsers.RecruitmentStockUser do
 
   schema "recruitment_stock_users" do
     field :skill_panel, :string
-    field :desired_income, :integer
+    field :desired_income, :decimal
     belongs_to :recruiter, User
     belongs_to :user, User
