+++ lib/bright/recruits/interview.ex
@@ -18,7 +18,7 @@ defmodule Bright.Recruits.Interview do
 
   schema "interviews" do
     field :skill_panel_name, :string
-    field :desired_income, :integer
+    field :desired_income, :decimal
     field :comment, :string
     field :skill_params, :string
     field :cancel_reason, :string
